// src/api/generated/models/transactionTypeResource.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Laravel
 * OpenAPI spec version: 0.0.1
 */

export interface TransactionTypeResource {
  id: string;
  name: string;
  /** @nullable */
  description: string | null;
  operation: string;
  created_at: string;
  updated_at: string;
}